mcqgen/requirements..txt
openai
langchain
streamlit
python-dotenv
pyPDF2

